Better handling of missing lock file with --use-locks=yes
The previous code was trying to go with the traditional "conda env update" route inside the exception handling for a missing lock file. This caused any errors during `create_update_env` to be chained to the `LockFileNotFoundError`, which is confusing.

This just moves the "conda env update" path outside the exception handling, so the exceptions no longer chain.