Fix platform selectors of included files when rendering lock files
The platform selectors (`[win]`, `[linux]`, etc) for files included via `includes:`
were not being honored when rendering lock files for all platforms.

The bug was that the `CondaPlatform` instance, which configures selectors
explicitly, was not being passed along to the function which handled the `includes` directive.

To make this error less likely to happen again, the `conda_platform` parameter is no longer optional.